BIS-1871: Cycle bug, hierarchy graphs

+++ server-application-server/source/java/ch/systemsx/cisd/openbis/generic/server/dataaccess/db/DataDAO.java
@@ -1202,7 +1202,7 @@ final class DataDAO extends AbstractGenericEntityWithPropertiesDAO<DataPE> imple
             Number childId = (Number) result[0];
             Number parentId = (Number) result[1];
 
-            Set<Long> parentIds = childIdToParentIdsMap.get(childId);
+            Set<Long> parentIds = childIdToParentIdsMap.get(childId.longValue());
             if (parentIds == null)
             {
                 parentIds = new HashSet<Long>();
